Move toolbar code into hook/helper
